+++ index.d.ts
@@ -123,6 +123,10 @@ declare namespace customPlugin {
   //
   // HTTP CLIENT
   //
+  type HttpClientMetrics = {
+    disabled: boolean
+    urlLabel: string
+  }
   interface HttpClientBaseOptions {
     headers?: http.IncomingHttpHeaders,
     timeout?: number,
@@ -131,11 +135,13 @@ declare namespace customPlugin {
     ca?: string,
     httpsAgent?: https.Agent,
     logger?: fastify.FastifyLoggerInstance,
-    isMiaHeaderInjected?: boolean
+    isMiaHeaderInjected?: boolean,
+    disableMetrics?: boolean
   }
   interface BaseHttpClientResponse {
     headers: http.IncomingHttpHeaders
     statusCode: number
+    duration: number
   }
   interface StreamResponse extends BaseHttpClientResponse {
     payload: NodeJS.ReadableStream
@@ -162,6 +168,7 @@ declare namespace customPlugin {
     errorMessageKey?: string;
     proxy?: HttpClientProxy;
     query?: Record<string, string>;
+    metrics?: HttpClientMetrics;
   }
 
   type RequestBody = any | Buffer | ReadableStream